+++ packages/interactive-wrapper/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.14](https://github.com/newsuk/times-components/compare/@times-components/interactive-wrapper@0.2.13...@times-components/interactive-wrapper@0.2.14) (2019-02-20)
+
+**Note:** Version bump only for package @times-components/interactive-wrapper
+
+
+
+
+
 ## [0.2.13](https://github.com/newsuk/times-components/compare/@times-components/interactive-wrapper@0.2.12...@times-components/interactive-wrapper@0.2.13) (2019-02-18)
 
 **Note:** Version bump only for package @times-components/interactive-wrapper