+++ packages/test-utils/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.2.4](https://github.com/newsuk/times-components/compare/@times-components/test-utils@2.2.3...@times-components/test-utils@2.2.4) (2019-02-20)
+
+**Note:** Version bump only for package @times-components/test-utils
+
+
+
+
+
 ## [2.2.3](https://github.com/newsuk/times-components/compare/@times-components/test-utils@2.2.2...@times-components/test-utils@2.2.3) (2019-02-18)
 
 **Note:** Version bump only for package @times-components/test-utils